mcp/requirements.txt
fastapi
uvicorn[standard]
optuna
pandas
requests
pynacl
python-dotenv
